server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==1.0.1
PyPDF2==3.0.1
aiofiles==23.2.1
python-jose==3.3.0
redis==5.0.1
pydantic==2.5.2
python-dotenv==1.0.0
nltk==3.8.1
spacy==3.7.2
# 新增的文件格式支持
openpyxl==3.1.2  # Excel文件
python-pptx==0.6.21  # PowerPoint文件
markdown2==2.4.10  # Markdown文件
beautifulsoup4==4.12.2  # HTML解析
lxml==4.9.3  # XML解析
ebooklib==0.18  # EPUB电子书
mobi==0.3.3  # Mobi电子书
odfpy==1.4.1  # OpenDocument格式
chardet==5.2.0  # 文件编码检测
pygments==2.17.2  # 代码文件语法高亮
